+++ BlazorShop.Data/Models/OrderProduct.cs
@@ -1,14 +1,12 @@
-﻿namespace BlazorShop.Data.Models
-{
+﻿namespace BlazorShop.Data.Models {
     using Contracts;
 
-    public class OrderProduct : BaseModel
-    {
+    public class OrderProduct : BaseModel {
         public string OrderId { get; set; }
 
         public Order Order { get; set; }
 
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
 
         public Product Product { get; set; }
