+++ BlazorShop.Data/Models/ShoppingCartProduct.cs
@@ -1,14 +1,12 @@
-﻿namespace BlazorShop.Data.Models
-{
+﻿namespace BlazorShop.Data.Models {
     using Contracts;
 
-    public class ShoppingCartProduct : BaseModel
-    {
-        public int ShoppingCartId { get; set; }
+    public class ShoppingCartProduct : BaseModel {
+        public long ShoppingCartId { get; set; }
 
         public ShoppingCart ShoppingCart { get; set; }
 
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
 
         public Product Product { get; set; }
