+++ BlazorShop.Models/Categories/CategoriesListingResponseModel.cs
@@ -1,11 +1,9 @@
-﻿namespace BlazorShop.Models.Categories
-{
+﻿namespace BlazorShop.Models.Categories {
     using Common.Mapping;
     using Data.Models;
 
-    public class CategoriesListingResponseModel : IMapFrom<Category>
-    {
-        public int Id { get; set; }
+    public class CategoriesListingResponseModel : IMapFrom<Category> {
+        public long Id { get; set; }
 
         public string Name { get; set; }
     }
