+++ BlazorShop.Models/Products/ProductsDetailsResponseModel.cs
@@ -1,10 +1,8 @@
-﻿namespace BlazorShop.Models.Products
-{
-    public class ProductsDetailsResponseModel : ProductsListingResponseModel
-    {
+﻿namespace BlazorShop.Models.Products {
+    public class ProductsDetailsResponseModel : ProductsListingResponseModel {
         public int Quantity { get; set; }
 
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
 
         public string CategoryName { get; set; }
     }
