+++ BlazorShop.Models/ShoppingCarts/ShoppingCartRequestModel.cs
@@ -1,13 +1,11 @@
-﻿namespace BlazorShop.Models.ShoppingCarts
-{
+﻿namespace BlazorShop.Models.ShoppingCarts {
     using System.ComponentModel.DataAnnotations;
 
     using static Data.ModelConstants.Product;
 
-    public class ShoppingCartRequestModel
-    {
+    public class ShoppingCartRequestModel {
         [Required]
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
 
         [Range(MinQuantity, MaxQuantity)]
         public int Quantity { get; set; } = MinQuantity;
