+++ BlazorShop.Services/Orders/OrdersService.cs
@@ -1,26 +1,20 @@
-﻿namespace BlazorShop.Services.Orders
-{
+﻿namespace BlazorShop.Services.Orders {
+    using AutoMapper;
+    using Data;
+    using Data.Models;
+    using Microsoft.EntityFrameworkCore;
+    using Models.Orders;
     using System.Collections.Generic;
     using System.Linq;
     using System.Threading.Tasks;
 
-    using AutoMapper;
-    using Microsoft.EntityFrameworkCore;
-
-    using Data;
-    using Data.Models;
-    using Models.Orders;
-
-    public class OrdersService : BaseService<Order>, IOrdersService
-    {
+    public class OrdersService : BaseService<Order>, IOrdersService {
         public OrdersService(BlazorShopDbContext data, IMapper mapper)
-            : base(data, mapper)
-        {
+            : base(data, mapper) {
         }
 
         public async Task<string> PurchaseAsync(
-            OrdersRequestModel model, string userId)
-        {
+            OrdersRequestModel model, string userId) {
             var order = new Order
             {
                 UserId = userId,
@@ -35,8 +29,7 @@
 
             var orderProducts = new List<OrderProduct>();
 
-            foreach (var shoppingCartProduct in shoppingCartProducts)
-            {
+            foreach (var shoppingCartProduct in shoppingCartProducts) {
                 var productId = shoppingCartProduct.ProductId;
                 var requestQuantity = shoppingCartProduct.Quantity;
 
@@ -68,8 +61,7 @@
                     .Where(o => o.Id == id))
                 .FirstOrDefaultAsync();
 
-        public async Task<IEnumerable<OrdersListingResponseModel>> ByUserAsync(
-            string userId)
+        public async Task<IEnumerable<OrdersListingResponseModel>> ByUserAsync(string userId)
             => await this.Mapper
                 .ProjectTo<OrdersListingResponseModel>(this
                     .AllAsNoTracking()
@@ -77,10 +69,7 @@
                     .SelectMany(o => o.Products))
                 .ToListAsync();
 
-        private async Task ReduceProductQuantity(
-            int productId,
-            int requestQuantity)
-        {
+        private async Task ReduceProductQuantity(long productId, int requestQuantity) {
             var product = await this
                 .Data
                 .Products
