+++ BlazorShop.Services/ShoppingCarts/IShoppingCartsService.cs
@@ -1,19 +1,16 @@
-﻿namespace BlazorShop.Services.ShoppingCarts
-{
-    using System.Collections.Generic;
-    using System.Threading.Tasks;
-
+﻿namespace BlazorShop.Services.ShoppingCarts {
     using Common;
     using Models;
     using Models.ShoppingCarts;
+    using System.Collections.Generic;
+    using System.Threading.Tasks;
 
-    public interface IShoppingCartsService : IService
-    {
+    public interface IShoppingCartsService : IService {
         Task<Result> AddProductAsync(ShoppingCartRequestModel model, string userId);
 
         Task<Result> UpdateProductAsync(ShoppingCartRequestModel model, string userId);
 
-        Task<Result> RemoveProductAsync(int productId, string userId);
+        Task<Result> RemoveProductAsync(long productId, string userId);
 
         Task<int> TotalAsync(string userId);
