+++ BlazorShop.Services/Wishlists/IWishlistsService.cs
@@ -1,17 +1,14 @@
-﻿namespace BlazorShop.Services.Wishlists
-{
-    using System.Collections.Generic;
-    using System.Threading.Tasks;
-
+﻿namespace BlazorShop.Services.Wishlists {
     using Common;
     using Models;
     using Models.Wishlists;
+    using System.Collections.Generic;
+    using System.Threading.Tasks;
 
-    public interface IWishlistsService : IService
-    {
-        Task<Result> AddProductAsync(int productId, string userId);
+    public interface IWishlistsService : IService {
+        Task<Result> AddProductAsync(long productId, string userId);
 
-        Task<Result> RemoveProductAsync(int productId, string userId);
+        Task<Result> RemoveProductAsync(long productId, string userId);
 
         Task<IEnumerable<WishlistsProductsResponseModel>> ByUserAsync(string userId);
     }
