+++ BlazorShop.Web/Client/Infrastructure/Services/Addresses/AddressesService.cs
@@ -1,16 +1,13 @@
-﻿namespace BlazorShop.Web.Client.Infrastructure.Services.Addresses
-{
+﻿namespace BlazorShop.Web.Client.Infrastructure.Services.Addresses {
+    using Extensions;
+    using Models;
+    using Models.Addresses;
     using System.Collections.Generic;
     using System.Net.Http;
     using System.Net.Http.Json;
     using System.Threading.Tasks;
 
-    using Extensions;
-    using Models;
-    using Models.Addresses;
-
-    public class AddressesService : IAddressesService
-    {
+    public class AddressesService : IAddressesService {
         private readonly HttpClient http;
 
         private const string AddressesPath = "api/addresses";
@@ -18,15 +15,14 @@
 
         public AddressesService(HttpClient http) => this.http = http;
 
-        public async Task<int> CreateAsync(AddressesRequestModel model)
-        {
+        public async Task<long> CreateAsync(AddressesRequestModel model) {
             var addressResponse = await this.http.PostAsJsonAsync(AddressesPath, model);
-            var addressId = await addressResponse.Content.ReadFromJsonAsync<int>();
+            var addressId = await addressResponse.Content.ReadFromJsonAsync<long>();
 
             return addressId;
         }
 
-        public async Task<Result> DeleteAsync(int id)
+        public async Task<Result> DeleteAsync(long id)
             => await this.http
                 .DeleteAsync(AddressesPathWithSlash + id)
                 .ToResult();
